后端完成，开始完成前端
前端完成了登录和注册

+++ src/main/java/entity/User.java
@@ -1,18 +1,18 @@
 package entity;
 
-
 import lombok.Data;
 
 @Data
 public class User {
 
-  private int userId;
+  private long userId;
   private String userName;
   private String userRname;
   private String gender;
-  private int birthYear;
+  private long birthYear;
   private String phone;
   private String idNumber;
   private String password;
   private String roleFlag;
+
 }
